feat: add useSearchbar hook for use with Searchbar component
